:construction: [refactor]: refactor conponents to .tsx with a lot of errors
